rename ddt4all.py to main.py
