--- DDT4ALL.BAT
@@ -1,2 +1,0 @@
-start Python27/Python.exe ddt4all.py
-pause
