+++ tools/po_create.sh
@@ -2,5 +2,5 @@
 
 xgettext elm.py -o locale/elm.pot --from-code=UTF-8
 xgettext parameters.py -o locale/ddt4all.pot --from-code=UTF-8
-xgettext ddt4all.py -o locale/ddt4all_main.pot --from-code=UTF-8
+xgettext main.py -o locale/ddt4all_main.pot --from-code=UTF-8
 xgettext dataeditor.py -o locale/dataeditor.pot --from-code=UTF-8
